- Added Parametervalidation

+++ src/NoisedCore/UserManagement/UserManager.cs
@@ -1,4 +1,5 @@
-﻿using Noised.Core.Crypto;
+﻿using System;
+using Noised.Core.Crypto;
 using Noised.Core.DB;
 using PasswordSecurity;
 
@@ -38,6 +39,11 @@ namespace Noised.Core.UserManagement
         
         public User CreateUser(string username, string password)
         {
+            if (string.IsNullOrWhiteSpace(username))
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(username));
+            if (string.IsNullOrWhiteSpace(password))
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(password));
+
             string hashedPw = _passwordManager.CreateHash(password);
         
             using (IUnitOfWork iuw = _dbFactory.GetUnitOfWork())
@@ -51,28 +57,35 @@ namespace Noised.Core.UserManagement
         
         public void DeleteUser(User user)
         {
+            if (user == null) throw new ArgumentNullException(nameof(user));
+
             using (IUnitOfWork iuw = _dbFactory.GetUnitOfWork())
             {
                 iuw.UserRepository.DeleteUser(user);
                 iuw.SaveChanges();
             }
         }
-        
+
         public void UpdateUser(User user)
         {
+            if (user == null) throw new ArgumentNullException(nameof(user));
+
             using (IUnitOfWork iuw = _dbFactory.GetUnitOfWork())
             {
                 iuw.UserRepository.UpdateUser(user);
                 iuw.SaveChanges();
             }
         }
-        
+
         public User GetUser(string username)
         {
+            if (string.IsNullOrWhiteSpace(username))
+                throw new ArgumentException("Value cannot be null or whitespace.", nameof(username));
+
             using (IUnitOfWork iuw = _dbFactory.GetUnitOfWork())
                 return iuw.UserRepository.GetUser(username);
         }
-        
+
         #endregion
     }
 }
